+++ realm/realm-library/src/main/cpp/io_realm_internal_OsResults.cpp
@@ -19,6 +19,7 @@
 #include <shared_realm.hpp>
 #include <results.hpp>
 #include <list.hpp>
+#include <realm/util/optional.hpp>
 
 #include "java_class_global_def.hpp"
 #include "java_sort_descriptor.hpp"
@@ -240,13 +241,15 @@ JNIEXPORT jlong JNICALL Java_io_realm_internal_OsResults_nativeDistinct(JNIEnv* 
 }
 
 JNIEXPORT void JNICALL Java_io_realm_internal_OsResults_nativeStartListening(JNIEnv* env, jobject instance,
-                                                                              jlong native_ptr)
+                                                                              jlong native_ptr, jstring j_subscription_name)
 {
     TR_ENTER_PTR(native_ptr)
 
     try {
         auto wrapper = reinterpret_cast<ResultsWrapper*>(native_ptr);
-        wrapper->start_listening(env, instance);
+        JStringAccessor subscription_name(env, j_subscription_name);
+        auto key = subscription_name.is_null_or_empty() ? util::none : util::Optional<std::string>(subscription_name);
+        wrapper->start_listening(env, instance, key);
     }
     CATCH_STD()
 }